rename cruciverbalists to cruciverbalist
