+++ benchmarks/direct.py
@@ -2,7 +2,7 @@ from contextlib import suppress
 
 from platyrhynchos.commons.logger import logger
 from platyrhynchos.crossword import CrosswordImprovable
-from platyrhynchos.cruciverbalists import EnglishSimpleCruciverbalist
+from platyrhynchos.cruciverbalist import EnglishSimpleCruciverbalist
 
 with suppress(ModuleNotFoundError):
     import functiontrace
